// app/(default)/thing/fleet/types/GetItemFleetsForClient.ts
export interface GetItemFleetsForClient_getItemFleetsForClient_page {
  __typename: "ItemFleetConnection";
  edges: GetItemFleetsForClient_getItemFleetsForClient_page_edges[] | null;
  pageInfo: GetItemFleetsForClient_getItemFleetsForClient_page_pageInfo | null;
}

export interface GetItemFleetsForClient_getItemFleetsForClient_pageData {
  __typename: "PageData";
  count: number;
  limit: number;
  offset: number;
}

export interface GetItemFleetsForClient_getItemFleetsForClient {
  __typename: "GetAllItemFleetsResponse";
  page: GetItemFleetsForClient_getItemFleetsForClient_page;
  pageData: GetItemFleetsForClient_getItemFleetsForClient_pageData | null;
}

export interface GetItemFleetsForClient {
  getItemFleetsForClient: GetItemFleetsForClient_getItemFleetsForClient;
}

export interface GetItemFleetsForClientVariables {
  after?: string | null;
  before?: string | null;
  clientId: string;
  first?: number | null;
  last?: number | null;
  search?: string | null;
}